apicast.conf: sort directives by execution order
The content phase is executed before the log one, but in the config,
'log_by_lua' appeared before 'content_by_lua'. There are some operations
performed to calculate run times and to understand them it's necessary to know
the order of execution of the phases. I think that sorting them in the config helps.