resty/http_ng: replace headers with normal table
The members of a 'headers' instance have more methods and make the merge
operation more costly. Also, we do not need to instantiate 'headers'
here, because it's done later anyway when calling http.request.new().
This also makes the code more readable in my opinion, because we only
need to deal with header instances in one class instead of in two like
we do now.

I have benchmarked this and it increases performance a bit when Apicast
acts both as auth backend and api backend (cpu workload).