cli/start: improve how the nginx binary is chosen
Now it also takes into account TEST_NGINX_BINARY, but
APICAST_OPENRESTY_BINARY takes precedence.
Also, .value is used instead of .get. .value evaluates to nil when an
env is "". This is useful for the or condition, as a value of "" should
be treated like it has not been set.