Add caching policy
This is based on the current CacheHandler module. Many parts of the code
are duplicated, but this duplication will disappear once we split the
current Apicast policy into smaller ones.

This does not break the current Apicast cache. Users can still use the
APICAST_BACKEND_CACHE_HANDLER to configure the cache.