policy/caching: use export to expose the cache handler
Using rewrite() instead of export() causes problems when combining this
policy with the apicast one.

The reason is that the apicast policy uses the cache handler in its
rewrite phase, and this policy sets it in the same phase. This means
that currently, the caching policy only works when placed before the
apicast policy in the chain. There's no reason why it should be that
way, and setting the handler in export() instead of rewrite() should fix
the issue.