policy/upstream: match rules in rewrite phase
This way this policy will be easier to combine with others, for example
the URL rewriting policy.

Before this change, the upstream policy ran on the content phase,
whereas the URL rewriting one ran on the rewrite phase. This means
that the upstream policy always took into account the rewritten path
instead of the original one, regardless of the position of the policies
in the chain. This commit fixes the issue.