[make] improve development environment
Use bound volumes to as cache. It is not possible to
both drop permissions of the docker container and use named volumes.
Dropping permissions is desired to not leave files owned by root in
the project folder. But named volumes are owned by root user and
are not group writeable. That makes them inacessible by non root users.

Moved the whole dev environment to docker-compose as there is no need
for building a container.