t/apicast-policy-3scale-batcher: change test strategy to make it closer to real usage
Before, the tests set a high reporting interval so a report was not
triggered in the middle of the test. The tests relied on calling a
specific endpoint to force a report to backend. Now, instead of doing
that, we let the policy to run report jobs as it would normally do, we
aggregate the data we receive in the reporting endpoint, and at the end
of the test, we verify it. This approach is closer to a real usage of
the policy.