configuration_loader/remote_v2: avoid sparse arrays of OIDC configs
APIcast crashes when loading a config where only some of the services use
OIDC.

The reason is that we created an array of OIDC configs with
size=number_of_services. Let's say we have 100 services and only the
50th has an OIDC config. In this case, we created this Lua table: { [50]
= oidc_config_here }.

The problem is that cjson raises an error when trying to convert a
sparse array like that into JSON.

Check
https://www.kyne.com.au/~mark/software/lua-cjson-manual.html#encode_sparse_array
for more details.

Assigning false instead of nil solves the issue.