t/apicast-policy-routing: delete redundant tests
I think that for this particular case it's enough to test operations
when they match. If we try to match a header and it works, it means that
the code was able to extract the header specified correctly, which is
one of the things we want to do in these integration tests. Adding an
extra test with the same information but evaluating to false does not
add much value.