policy/apicast: add flag to run post_action() only if acess() ran
This commit adds a flag to run post_action() only when access() was
executed.

Other policies can call ngx.exit(4xx) on access() or rewrite. If they're
placed before APIcast in the chain, the request will be denied and this
access() phase will no be run. However, ngx.exit(4xx) skips some phases,
but not post_action. Post_action would run if we did not set this flag,
and we want to avoid that. Otherwise, post_action could call authrep()
even when another policy denied the request.