policy_pusher: send access token correctly
Notice that there was a 'token:url' in the code and the tests instead of
'token@url'. While at it, I've changed this to send it via params
because that makes the tests a bit easier (do not need to convert things
to base64) and it's closer to what the system UI shows because
'access_token' appears as a param there.